Split up the ingests API tests for easier readability
